requirements.txt
colorlog==6.7.0
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.5
jwcrypto==1.4.2
gunicorn==20.1.0
PyJWT==2.4.0
python-jwt==3.3.2
pyotp==2.6.0
